requirements/base.txt
attrs==25.3.0
cffi==1.17.1
cryptography==44.0.2
dill==0.3.9
ecdsa==0.19.1
inquirerpy==0.3.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
multiprocess==0.70.14
orjson==3.9.15
pfzy==0.3.4
pony==0.7.19
prompt_toolkit==3.0.50
pyasn1==0.4.8
pycparser==2.22
python-dotenv==1.1.0
python-jose==3.4.0
referencing==0.36.2
robyn==0.65.0
rpds-py==0.24.0
rsa==4.9
rustimport==1.3.4
six==1.17.0
toml==0.10.2
typing_extensions==4.13.0
uvloop==0.19.0
watchdog==4.0.1
wcwidth==0.2.13
Werkzeug==3.1.3
